+++ examples/get-started/bundle-with-vite/vite.config.ts
@@ -12,9 +12,8 @@ const getAliases = async (frameworkName, frameworkRootDir) => {
   return aliases
 }
 
-// https://vitejs.dev/config/
+/** https://vitejs.dev/config/ */
 export default defineConfig(async () => ({
   resolve: {alias: await getAliases('@loaders.gl', `${__dirname}/../../..`)},
   server: {open: true}
 }))
-