+++ modules/3d-tiles/src/lib/parsers/helpers/parse-3d-implicit-tiles.ts
@@ -197,11 +197,7 @@ function formatTileData(
     root = {region: [...region, s2bv.minimumHeight, s2bv.maximumHeight]};
   }
 
-  const boundingVolume = calculateBoundingVolumeForChildTile(
-    level,
-    root,
-    childCoordinates
-  );
+  const boundingVolume = calculateBoundingVolumeForChildTile(level, root, childCoordinates);
 
   return {
     children: tile.children,