+++ modules/core/src/iterators/make-stream/make-node-stream.ts
@@ -1,4 +1,10 @@
-import {Readable, ReadableOptions} from 'stream';
+import type {ReadableOptions} from 'stream';
+import * as Stream from 'stream';
+
+class _Readable {}
+
+type ReadableType = Stream.Readable | _Readable;
+const Readable = Stream.Readable || _Readable;
 
 export type MakeStreamOptions = ReadableOptions;
 
@@ -6,7 +12,7 @@ export type MakeStreamOptions = ReadableOptions;
 export function makeStream<ArrayBuffer>(
   source: Iterable<ArrayBuffer> | AsyncIterable<ArrayBuffer>,
   options?: ReadableOptions
-): Readable {
+): ReadableType {
   const iterator = source[Symbol.asyncIterator]
     ? source[Symbol.asyncIterator]()
     : source[Symbol.iterator]();