+++ modules/wms/test/index.ts
@@ -11,3 +11,9 @@ import './wmts/wmts-capabilities-loader.spec';
 import './wfs/wfs-capabilities-loader.spec';
 
 import './gml/gml-loader.spec';
+
+import './lerc/lerc-sanity.spec';
+import './lerc/lerc-level2.spec';
+
+// Services
+import './data-sources/wms-service.spec';